Corrected controlling of tesseract post route and validation of schema.
Also adding functionality to when a valid post request to tesseract
is received a url query string is returned.

Set todo's for models, validtion schemas, controller logic, and class
representation of data. See todo's for further explanation of what needs
to be completed.

Instead of using handwritten file to validate schema we are now using the
npm package called celebrate. See tesseract_controller.js for an example.